+++ checks/store.go
@@ -1,8 +1,6 @@
 package checks
 
 import (
-	"my/checker/store"
-
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
