+++ store/mongodb/functions.go
@@ -1,0 +1,14 @@
+package store
+
+func InitDB(dbConfig DBConfig) (IStore, error) {
+	switch dbConfig.Protocol {
+	case "mongodb":
+		client := &mongoDbStore{}
+		store, err := client.Init(dbConfig)
+		if err != nil {
+			return nil, err
+		}
+		Store = store
+	}
+	return Store, nil
+}
