+++ store/mongodb/interfaces.go
@@ -1,0 +1,6 @@
+package store
+
+type IStore interface {
+	Init(dbConfig DBConfig) (IStore, error)
+	Disconnect() error
+}
