+++ store/mongodb/main.go
@@ -2,12 +2,11 @@ package store
 
 import (
 	"context"
-	"my/checker/config"
 )
 
 var (
-	logger     = config.GetLog()
-	configurer = config.GetConfig()
+	// logger     = config.GetLog()
+	// configurer = config.GetConfig()
 
 	Store     IStore
 	DBContext context.Context
